iot_app/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pyjwt
passlib[bcrypt]
paho-mqtt
jinja2
python-multipart
